# openapi.yaml
openapi: 3.0.0
info:
  title: betterHR Microservices API
  description: API for Identity, Employee Profiles, Attendance, and Leave Management.
  version: 2.0.0
servers:
  - url: https://api.betterhr.com/prod
...
components:
  securitySchemes:
    CognitoAuth:
      type: apiKey
      in: header
      name: Authorization
paths:
...
  /employees:
    get:
      tags: [Employee Profile]
      summary: List employees
      security: [{ CognitoAuth: [] }]
      responses: { '200': { description: OK } }
    post:
      tags: [Employee Profile]
      summary: Onboard employee
      security: [{ CognitoAuth: [] }]
      responses: { '201': { description: Created } }
  /employees/{id}:
    get:
      tags: [Employee Profile]
      summary: Get employee profile
      security: [{ CognitoAuth: [] }]
      responses: { '200': { description: OK } }
...